refactor: 주문/취소 후처리 Kafka 이벤트 기반 구조로 통합 (#115)
- OrderCancelledEvent 신규 추가 (deal-core)
- OrderService/Controller/ServiceImpl: Query/Async 분리 클래스 통합
- CheckoutServiceImpl: OrderCompletedEvent Spring Event 발행으로 후처리 분리
- OrderServiceImpl.cancelOrder: OrderCancelledEvent Spring Event 발행
- OrderCancelledEventPublisher: AFTER_COMMIT → order-cancelled-events Kafka 발행
- deal-api-user EmailService/Impl 제거 (deal-consumer로 일원화)
- deal-consumer: OrderCancelEmailConsumer / OrderCancelNotificationConsumer / OrderCancelStatsConsumer 신규
- deal-api-user build.gradle.kts: spring-boot-starter-mail 의존성 제거
- OrderConfirmService: Payment.paidAt 필드 추가
