+++ deal-api-user/build.gradle.kts
@@ -2,5 +2,4 @@ dependencies {
     implementation(project(":deal-common-api"))
     implementation("org.springframework.boot:spring-boot-starter-validation")
     implementation("org.apache.tika:tika-core:2.9.1")
-    implementation("org.springframework.boot:spring-boot-starter-mail")
 }
