+++ deal-api-user/src/main/kotlin/com/chaltteok/user/checkout/service/CheckoutServiceImpl.kt
@@ -1,15 +1,12 @@
 package com.chaltteok.user.checkout.service
 
 import com.chaltteok.common.exception.BusinessException
-import com.chaltteok.core.domain.Notification
 import com.chaltteok.core.domain.Order
-import com.chaltteok.core.domain.enums.NotificationType
 import com.chaltteok.core.domain.OrderItem
-import com.chaltteok.core.domain.enums.OrderStatus
 import com.chaltteok.core.domain.Payment
+import com.chaltteok.core.domain.enums.OrderStatus
 import com.chaltteok.core.domain.enums.PaymentStatus
 import com.chaltteok.core.event.OrderCompletedEvent
-import com.chaltteok.core.repository.notification.NotificationRepository
 import com.chaltteok.core.repository.order.OrderRepository
 import com.chaltteok.core.repository.orderitem.OrderItemRepository
 import com.chaltteok.core.repository.payment.PaymentRepository
@@ -18,15 +15,16 @@ import com.chaltteok.core.repository.user.UserRepository
 import com.chaltteok.user.checkout.dto.CheckoutRequest
 import com.chaltteok.user.checkout.dto.CheckoutResponse
 import com.chaltteok.user.checkout.enums.CheckoutErrorCode
-import com.chaltteok.core.service.orderstats.OrderStatsService
-import com.chaltteok.user.order.service.EmailService
 import io.github.oshai.kotlinlogging.KotlinLogging
+import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
-import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
 
 private val logger = KotlinLogging.logger {}
+private val ORDER_DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")
+
 @Service
 class CheckoutServiceImpl(
     private val userRepository: UserRepository,
@@ -34,9 +32,7 @@ class CheckoutServiceImpl(
     private val orderRepository: OrderRepository,
     private val orderItemRepository: OrderItemRepository,
     private val paymentRepository: PaymentRepository,
-    private val notificationRepository: NotificationRepository,
-    private val orderStatsService: OrderStatsService,
-    private val emailService: EmailService
+    private val applicationEventPublisher: ApplicationEventPublisher,
 ) : CheckoutService {
 
     @Transactional
@@ -67,39 +63,39 @@ class CheckoutServiceImpl(
         }
         orderItemRepository.saveAll(orderItems)
 
-        val payment = Payment(order = savedOrder, amount = request.totalAmount.toInt(), status = PaymentStatus.SUCCESS, paymentMethod = request.paymentMethod)
-        paymentRepository.save(payment)
-
-        savedOrder.status = OrderStatus.COMPLETED
-        val start = System.currentTimeMillis()
-        emailService.sendOrderConfirmation(OrderCompletedEvent(
-            orderId = savedOrder.id ?: error("Order ID가 저장 후에도 null입니다"),
-            orderNumber = order.orderNumber,
-            userEmail = user.email,
-            userName = user.nickname,
-            productName = NotificationType.ORDER.name,
-            totalAmount = savedOrder.totalPrice.toLong(),
-            orderedAt = order.orderedAt.format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm")),
-        ))
-        orderStatsService.incrementOrderStats(
-            date = LocalDate.now(),
-            revenue = request.totalAmount,
-        )
-
-        val productNames = orderItems.joinToString(", ") { it.product.name }
-            .let { if (it.length > 450) it.take(450) + "…" else it }
-        notificationRepository.save(
-            Notification(
-                type = NotificationType.ORDER.name,
-                title = "새 주문이 들어왔습니다",
-                message = "$productNames (%,d원)".format(request.totalAmount),
+        paymentRepository.save(
+            Payment(
+                order = savedOrder,
+                amount = request.totalAmount.toInt(),
+                status = PaymentStatus.SUCCESS,
+                paymentMethod = request.paymentMethod,
+                paidAt = LocalDateTime.now(),
             )
         )
 
-        logger.info("부가 처리 시간: ${System.currentTimeMillis() - start}ms")
+        savedOrder.status = OrderStatus.COMPLETED
+
+        val orderId = savedOrder.id ?: error("Order ID가 저장 후에도 null입니다")
+        val productName = orderItems.joinToString(", ") { it.product.name }
+            .let { if (it.length > 450) it.take(450) + "…" else it }
+
+        // 트랜잭션 커밋 후 이메일·알림·통계를 각 EventListener가 처리
+        applicationEventPublisher.publishEvent(
+            OrderCompletedEvent(
+                orderId = orderId,
+                orderNumber = savedOrder.orderNumber,
+                userEmail = user.email,
+                userName = user.nickname,
+                productName = productName,
+                totalAmount = savedOrder.totalPrice.toLong(),
+                orderedAt = savedOrder.orderedAt.format(ORDER_DATE_FORMATTER),
+            )
+        )
+
+        logger.info { "일반 주문 완료 — orderNumber=${savedOrder.orderNumber}" }
 
         return CheckoutResponse(
-            orderId = savedOrder.id!!,
+            orderId = orderId,
             totalAmount = savedOrder.totalPrice.toLong(),
             status = savedOrder.status.name,
         )
