--- deal-api-user/src/main/kotlin/com/chaltteok/user/order/service/EmailService.kt
@@ -1,7 +1,0 @@
-package com.chaltteok.user.order.service
-
-import com.chaltteok.core.event.OrderCompletedEvent
-
-interface EmailService {
-    fun sendOrderConfirmation(event: OrderCompletedEvent)
-}
