--- deal-api-user/src/main/kotlin/com/chaltteok/user/order/service/OrderQueryService.kt
@@ -1,18 +1,0 @@
-package com.chaltteok.user.order.service
-
-import com.chaltteok.user.order.dto.OrderHistoryPageResponse
-import com.chaltteok.user.order.dto.OrderHistoryResponse
-import org.springframework.data.domain.Pageable
-
-interface OrderQueryService {
-    fun getOrderHistory(
-        userId: Long,
-        keyword: String?,
-        status: String?,
-        fromDate: String?,
-        toDate: String?,
-        paymentStatus: String?,
-        pageable: Pageable,
-    ): OrderHistoryPageResponse
-    fun getOrderDetail(userId: Long, orderNumber: String): OrderHistoryResponse
-}
