+++ deal-core/src/main/kotlin/com/chaltteok/core/event/OrderCancelledEvent.kt
@@ -1,0 +1,10 @@
+package com.chaltteok.core.event
+
+data class OrderCancelledEvent(
+    val orderId: Long,
+    val orderNumber: String,
+    val userEmail: String,
+    val userName: String,
+    val totalAmount: Long,
+    val cancelledAt: String,
+)
